[Refactor] python structure refactor
