+++ mailtracking/constants.py
